controllers-films-users Исправить замечания ревьювера 2
