+++ .github/workflows/api-tests.yml
@@ -1,0 +1,8 @@
+name: Films API Tests
+
+on:
+  pull_request:
+
+jobs:
+  build:
+    uses: yandex-praktikum/java-filmorate/.github/workflows/api-tests.yml@ci
