+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -48,7 +48,7 @@ public class UserController {
             log.error("Электронная почта не может быть пустой и должна содержать символ @!");
             throw new ValidationException("Электронная почта не может быть пустой и должна содержать символ @!");
         }
-        if (user.getLogin() == null || user.getLogin().isBlank() || user.getLogin().contains(" ")) {
+        if (user.getLogin() == null || user.getLogin().isBlank()) {
             log.error("Логин не может быть пустым и содержать пробелы!");
             throw new ValidationException("Логин не может быть пустым и содержать пробелы!");
         }
@@ -56,7 +56,7 @@ public class UserController {
             log.warn("Имя для отображения может быть пустым — в таком случае будет использован логин!");
             user.setName(user.getLogin());
         }
-        if (user.getBirthday() != null & user.getBirthday().isAfter(LocalDate.now())) {
+        if (user.getBirthday() != null && user.getBirthday().isAfter(LocalDate.now())) {
             log.error("Дата рождения не может быть в будущем!");
             throw new ValidationException("Дата рождения не может быть в будущем!");
         }
